backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pyyaml==6.0.1
python-dotenv==1.0.0
PyJWT==2.8.0
